GO-P-2026-035 (legal retrieval): CONFIRM -> honest negative on subtle flip, robust anti
Held-out disjoint opinions (pairs_eval, n=1293), frozen base_bits=0.4:
  recon-trade  recon(O)=0.218 <= R=0.260   PASS
  anti worst   200/200 (A=0.612)           PASS
  flip         0/200                        FAIL
  -> NOT CONFIRMED. On held-out, recon-optimal BEATS read-preserving on AUROC
  (O=0.773 vs R=0.757), reversing calibration's razor-thin R>O (0.746 vs 0.742).

Diagnosis: the estimated read operator (centered covariance, fit on calibration)
overfit and did not transfer to disjoint opinions; faithful reconstruction
generalized better. The calibration flip (168/200) was over-optimistic selection
on a ~0.004 AUROC gap within noise. Differs from confirmed GO-P-2026-009 (planted
recon-matched probes, known read subspace) -- an ESTIMATED subspace is transfer-
fragile. The flip needs a read operator that GENERALIZES, not just read/signal
misalignment. GO-032/D4 shape (robust anti, no subtle flip). No file drawer.

Also: arg-parse fix (float base_bits); does not affect the sealed config.

+++ experiments/legal_flip.py
@@ -131,7 +131,7 @@ def main():
                   f"| flip {r['flip_boot']}/{r['boots']} | anti {r['anti_boot']}/{r['boots']} | recon-trade {r['recon_trade']}")
         return
     if "--confirm" in sys.argv:
-        bb = int(sys.argv[sys.argv.index("--confirm") + 1])
+        bb = float(sys.argv[sys.argv.index("--confirm") + 1])
         EAh, EBh = embed_cache(HELD_PAIRS, 1300, "held")      # DISJOINT opinions
         r = evaluate(EAh, EBh, mu, V, sig2, bb)
         f = lambda a, b: a / b if b else 0
